backend/requirements.txt
fastapi
uvicorn[standard]
supabase
openai
groq
google-genai
playwright
folium
python-dotenv
pydantic
pydantic-settings
httpx
jinja2
python-multipart
reportlab
pandas
openpyxl
psutil
sentry-sdk[fastapi]
